+++ src/components/Synthetics/StatusNotification/StatusNotification.scss
